Add ferry graphql dependency to the frontend - ferry graphql alow us to easily communicate with the backend graphql api
