+++ junqo_front/README.md
@@ -1,16 +1,39 @@
+<!-- omit in toc -->
 # junqo_front
 
-A new Flutter project.
+The mobile and web application for Junqo.
+
+<!-- omit in toc -->
+## Table of Contents
+
+- [Getting Started](#getting-started)
+- [Setup](#setup)
+- [Run](#run)
 
 ## Getting Started
-
-This project is a starting point for a Flutter application.
-
-A few resources to get you started if this is your first Flutter project:
-
-- [Lab: Write your first Flutter app](https://docs.flutter.dev/get-started/codelab)
-- [Cookbook: Useful Flutter samples](https://docs.flutter.dev/cookbook)
 
 For help getting started with Flutter development, view the
 [online documentation](https://docs.flutter.dev/), which offers tutorials,
 samples, guidance on mobile development, and a full API reference.
+
+## Setup
+
+Get dependencies
+
+```sh
+flutter pub get packages
+```
+
+Build graphql documents
+
+```sh
+dart run build_runner build
+```
+
+## Run
+
+Run the application
+
+```sh
+flutter run
+```
