+++ junqo_front/lib/graphql/schema.graphql
@@ -1,0 +1,73 @@
+# Junqo Project
+# This file contains the schema for the Junqo API
+
+# Endpoints
+type Query {
+    # Users
+    users: [User!]!
+    user(id: ID!): User!
+
+    # Messages
+    messages: [Message!]!
+    message(id: ID!): Message!
+
+    # Offers
+    offers: [Offer!]!
+    offer(id: ID!): Offer!
+
+    # AI
+    profileFromResume(resume: String!): User!
+    interviewReply(interview: String!): String!
+}
+
+type Mutation {
+    # Users
+    createUser(name: String!, email: String!, password: String!): User!
+    updateUser(id: ID!, name: String, email: String, password: String): User!
+    deleteUser(id: ID!): User!
+
+    # Messages
+    createMessage(text: String!): Message!
+    updateMessage(id: ID!, text: String): Message!
+    deleteMessage(id: ID!): Message!
+
+    # Offers
+    createOffer(title: String!, description: String!): Offer!
+    updateOffer(id: ID!, title: String, description: String): Offer!
+    deleteOffer(id: ID!): Offer!
+}
+
+type Subscription {
+    # Messages
+    messageAdded: Message!
+}
+
+# Types
+## Users
+type User {
+    id: ID!
+    name: String!
+    email: String!
+    password: String!
+    createdAt: String!
+    updatedAt: String!
+}
+
+## Messages
+type Message {
+    id: ID!
+    text: String!
+    createdAt: String!
+    updatedAt: String!
+    user: User!
+}
+
+## Offers
+type Offer {
+    id: ID!
+    title: String!
+    description: String!
+    createdAt: String!
+    updatedAt: String!
+    user: User!
+}
