+++ junqo_front/lib/src/client.dart
@@ -1,0 +1,27 @@
+import 'package:gql_http_link/gql_http_link.dart';
+import 'package:ferry/ferry.dart';
+import 'package:ferry_hive_store/ferry_hive_store.dart';
+import 'package:hive_flutter/adapters.dart';
+
+const apiUrl = 'localhost:3000';
+
+Future<Client> initClient() async {
+  await Hive.initFlutter();
+
+  final box = await Hive.openBox<Map<String, dynamic>>("graphql");
+
+  await box.clear();
+
+  final store = HiveStore(box);
+
+  final cache = Cache(store: store);
+
+  final link = HttpLink(apiUrl);
+
+  final client = Client(
+    link: link,
+    cache: cache,
+  );
+
+  return client;
+}
